initial consolidation into AppTask and manager file
