+++ examples/oven-app/silabs/BUILD.gn
@@ -105,10 +105,7 @@ silabs_executable("oven_app") {
   sources = [
     "${example_oven_dir}/oven-app-common/src/CookEndpoints.cpp",
     "${example_oven_dir}/oven-app-common/src/OvenEndpoint.cpp",
-    "src/AppSupportedTemperatureLevelsDelegate.cpp",
     "src/AppTask.cpp",
-    "src/DataModelCallbacks.cpp",
-    "src/OvenBindingHandler.cpp",
     "src/OvenManager.cpp",
   ]
   if (!disable_lcd) {
